refactor: kebab-case file names for consistency
apps/web (and packages/ui) use kebab-case filenames everywhere; a few
PascalCase/camelCase stragglers slipped in. Rename them to match:

- EditorView.tsx          -> editor-view.tsx
- ForkEditorView.tsx      -> fork-editor-view.tsx
- lib/utils/colorUtils.ts -> lib/utils/color-utils.ts
- ui .../colorUtils.ts    -> .../color-utils.ts (update raised-button import)

Component export names are unchanged (still PascalCase); only file names move.
