+++ apps/web/app/component/[id]/edit/editor-view.tsx
